Refactor to standalone package
