+++ pkg/orchestrator/gitHubActions.go
@@ -1,0 +1,22 @@
+package orchestrator
+
+import "os"
+
+type GitHubActionsConfigProvider struct{}
+
+func (a *GitHubActionsConfigProvider) GetBranchBuildConfig() BranchBuildConfig {
+	return BranchBuildConfig{Branch: os.Getenv("GITHUB_REF")}
+}
+
+func (a *GitHubActionsConfigProvider) GetPullRequestConfig() PullRequestConfig {
+	return PullRequestConfig{
+		Branch: os.Getenv("GITHUB_HEAD_REF"),
+		Base:   os.Getenv("GITHUB_BASE_REF"),
+		Key:    os.Getenv("GITHUB_EVENT_PULL_REQUEST_NUMBER"),
+	}
+}
+
+func (a *GitHubActionsConfigProvider) IsPullRequest() bool {
+	_, exists := os.LookupEnv("GITHUB_HEAD_REF")
+	return exists
+}
